Fix a bug in sqrt
We're doing a binary search for the square root between 1 and the given
number, and we test each number by squaring it.  If the starting number
is too large, though, we can end up with an overflow, so bound the
high mark to the largest number a square root can possibly be.